# swaggerthing.json
{
...
  "swagger": "2.0",
...
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "schemes": ["http"],
  "paths": {
    "/Buildings/Locale/{localeId}": {
      "get": {
        "tags": ["Buildings"],
        "operationId": "Buildings_GetBuildingsFromLocale",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "string",
            "name": "localeId",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "Page",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "PageSize",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "Filter",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "OrderBy",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "Include",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "array",
            "name": "IncludeProperties",
            "in": "query",
            "collectionFormat": "multi",
            "x-nullable": true,
            "items": {
              "type": "string"
            }
          },
          {
            "type": "string",
            "name": "Expand",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "array",
            "name": "ExpandProperties",
            "in": "query",
            "collectionFormat": "multi",
            "x-nullable": true,
            "items": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Building"
              }
            }
          }
        }
      }
    },
    "/Buildings/{id}": {
      "get": {
        "tags": ["Buildings"],
        "operationId": "Buildings_GetBuildingById",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "include",
            "in": "query",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Building"
            }
          }
        }
      },
      "patch": {
        "tags": ["Buildings"],
        "operationId": "Buildings_PatchBuilding",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "name": "patchDoc",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JsonPatchDocumentOfBuilding"
            },
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "include",
            "in": "query",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Building"
            }
          }
        }
      },
      "put": {
        "tags": ["Buildings"],
        "operationId": "Buildings_PutBuilding",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "name": "building",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Building"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": ["Buildings"],
        "operationId": "Buildings_DeleteBuilding",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Building"
            }
          }
        }
      }
    },
    "/Buildings": {
      "post": {
        "tags": ["Buildings"],
        "operationId": "Buildings_PostBuilding",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "name": "building",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Building"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Building"
            }
          }
        }
      }
    },
...
    "/Monsters": {
      "get": {
        "tags": ["Monsters"],
        "operationId": "Monsters_GetMonsters",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "integer",
            "name": "Page",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "PageSize",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "Filter",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "OrderBy",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "Include",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "array",
            "name": "IncludeProperties",
            "in": "query",
            "collectionFormat": "multi",
            "x-nullable": true,
            "items": {
              "type": "string"
            }
          },
          {
            "type": "string",
            "name": "Expand",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "array",
            "name": "ExpandProperties",
            "in": "query",
            "collectionFormat": "multi",
            "x-nullable": true,
            "items": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Monster"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Monsters"],
        "operationId": "Monsters_PostMonster",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "name": "monster",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Monster"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Monster"
            }
          }
        }
      }
    },
    "/Monsters/{id}": {
      "get": {
        "tags": ["Monsters"],
        "operationId": "Monsters_GetMonsterById",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "include",
            "in": "query",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Monster"
            }
          }
        }
      },
      "patch": {
        "tags": ["Monsters"],
        "operationId": "Monsters_PatchMonster",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "name": "patchDoc",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JsonPatchDocumentOfMonster"
            },
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "include",
            "in": "query",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Monster"
            }
          }
        }
      },
      "put": {
        "tags": ["Monsters"],
        "operationId": "Monsters_PutMonster",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "name": "monster",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Monster"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": ["Monsters"],
        "operationId": "Monsters_DeleteMonster",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Monster"
            }
          }
        }
      }
    },
...
    "/Npcs": {
      "get": {
        "tags": ["Npcs"],
        "operationId": "Npcs_GetAll",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "integer",
            "name": "Page",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "PageSize",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "Filter",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "OrderBy",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "Include",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "array",
            "name": "IncludeProperties",
            "in": "query",
            "collectionFormat": "multi",
            "x-nullable": true,
            "items": {
              "type": "string"
            }
          },
          {
            "type": "string",
            "name": "Expand",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "array",
            "name": "ExpandProperties",
            "in": "query",
            "collectionFormat": "multi",
            "x-nullable": true,
            "items": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Npc"
              }
            }
          }
        }
      }
    },
    "/Npcs/{id}": {
      "get": {
        "tags": ["Npcs"],
        "operationId": "Npcs_Get",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "include",
            "in": "query",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Npc"
            }
          }
        }
      },
      "patch": {
        "tags": ["Npcs"],
        "operationId": "Npcs_Patch",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "name": "patchDoc",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JsonPatchDocumentOfNpc"
            },
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "include",
            "in": "query",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Npc"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Building": {
      "allOf": [
        {
          "$ref": "#/definitions/Base"
        },
        {
          "type": "object",
          "properties": {
            "localeId": {
              "type": "string",
              "format": "guid"
            },
            "locale": {
              "$ref": "#/definitions/Locale"
            },
            "map": {
              "type": "string"
            },
            "npcs": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Npc"
              }
            },
            "monsters": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/MonsterBuilding"
              }
            },
            "players": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Player"
              }
            },
            "maps": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BuildingMap"
              }
            }
          }
        }
      ]
    },
    "Locale": {
      "allOf": [
        {
          "$ref": "#/definitions/Base"
        },
        {
          "type": "object",
          "properties": {
            "regionId": {
              "type": "string",
              "format": "guid"
            },
            "region": {
              "$ref": "#/definitions/Region"
            },
            "buildings": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Building"
              }
            },
            "dungeons": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Dungeon"
              }
            },
            "players": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Player"
              }
            },
            "npcs": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Npc"
              }
            },
            "monsters": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/MonsterLocale"
              }
            },
            "maps": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Map"
              }
            }
          }
        }
      ]
    },
    "Region": {
      "allOf": [
        {
          "$ref": "#/definitions/Base"
        },
        {
          "type": "object",
          "properties": {
            "locales": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Locale"
              }
            },
            "continentId": {
              "type": "string",
              "format": "guid"
            },
            "continent": {
              "$ref": "#/definitions/Continent"
            },
            "map": {
              "type": "string"
            }
          }
        }
      ]
    },
...
    "Base": {
      "type": "object",
      "required": ["id"],
      "properties": {
        "id": {
          "type": "string",
          "format": "guid"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "Player": {
      "allOf": [
        {
          "$ref": "#/definitions/Creature"
        },
        {
          "type": "object",
          "required": ["level", "xp", "inspiration"],
          "properties": {
            "level": {
              "type": "integer",
              "format": "int32"
            },
            "xp": {
              "type": "integer",
              "format": "int32"
            },
            "inspiration": {
              "type": "boolean"
            },
            "playerName": {
              "type": "string"
            },
            "background": {
              "type": "string"
            },
            "faction": {
              "type": "string"
            },
            "race": {
              "type": "string"
            },
            "localeId": {
              "type": "string",
              "format": "guid"
            },
            "locale": {
              "$ref": "#/definitions/Locale"
            },
            "buildingId": {
              "type": "string",
              "format": "guid"
            },
            "building": {
              "$ref": "#/definitions/Building"
            }
          }
        }
      ]
    },
    "Creature": {
      "allOf": [
        {
          "$ref": "#/definitions/Base"
        },
        {
          "type": "object",
          "required": [
            "strength",
            "dexterity",
            "constitution",
            "intelligence",
            "wisdom",
            "charisma",
            "armorClass",
            "hitPoints",
            "alignment"
          ],
          "properties": {
            "strength": {
              "type": "integer",
              "format": "int32"
            },
            "dexterity": {
              "type": "integer",
              "format": "int32"
            },
            "constitution": {
              "type": "integer",
              "format": "int32"
            },
            "intelligence": {
              "type": "integer",
              "format": "int32"
            },
            "wisdom": {
              "type": "integer",
              "format": "int32"
            },
            "charisma": {
              "type": "integer",
              "format": "int32"
            },
            "proficiencies": {
              "type": "array",
              "items": {}
            },
            "armorClass": {
              "type": "integer",
              "format": "int32"
            },
            "hitPoints": {
              "type": "integer",
              "format": "int32"
            },
            "hitDice": {
              "type": "string"
            },
            "size": {
              "type": "string"
            },
            "speed": {
              "type": "array",
              "items": {}
            },
            "languages": {
              "type": "string"
            },
            "alignment": {
              "$ref": "#/definitions/Alignment"
            },
            "reactions": {
              "type": "array",
              "items": {}
            },
            "picture": {
              "type": "string"
            }
          }
        }
      ]
    },
...
    "Npc": {
      "allOf": [
        {
          "$ref": "#/definitions/Base"
        },
        {
          "type": "object",
          "properties": {
            "picture": {
              "type": "string"
            },
            "background": {
              "type": "string"
            },
            "noteableEvents": {
              "type": "array",
              "items": {}
            },
            "beliefs": {
              "type": "array",
              "items": {}
            },
            "passions": {
              "type": "array",
              "items": {}
            },
            "flaws": {
              "type": "array",
              "items": {}
            },
            "monsterId": {
              "type": "string",
              "format": "guid"
            },
            "monster": {
              "$ref": "#/definitions/Monster"
            },
            "localeId": {
              "type": "string",
              "format": "guid"
            },
            "locale": {
              "$ref": "#/definitions/Locale"
            },
            "buildingId": {
              "type": "string",
              "format": "guid"
            },
            "building": {
              "$ref": "#/definitions/Building"
            }
          }
        }
      ]
    },
    "Monster": {
      "allOf": [
        {
          "$ref": "#/definitions/Creature"
        },
        {
          "type": "object",
          "required": [
            "challengeRating",
            "xp",
            "passivePerception",
            "monsterType"
          ],
          "properties": {
            "challengeRating": {
              "type": "number",
              "format": "double"
            },
            "xp": {
              "type": "integer",
              "format": "int32"
            },
            "passivePerception": {
              "type": "integer",
              "format": "int32"
            },
            "monsterType": {
              "$ref": "#/definitions/MonsterType"
            },
            "actions": {
              "type": "array",
              "items": {}
            },
            "legendaryActions": {
              "type": "array",
              "items": {}
            },
            "specialAbilities": {
              "type": "array",
              "items": {}
            },
            "senses": {},
            "npcs": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Npc"
              }
            },
            "locales": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/MonsterLocale"
              }
            },
            "buildings": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/MonsterBuilding"
              }
            }
          }
        }
      ]
    },
    "MonsterType": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "Abberation",
        "Beast",
        "Celestial",
        "Construct",
        "Dragon",
        "Elemental",
        "Fey",
        "Fiend",
        "Giant",
        "Humanoid",
        "Monstrosity",
        "Ooze",
        "Plant",
        "Undead",
        "None"
      ],
      "enum": [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14]
    },
    "MonsterLocale": {
      "type": "object",
      "required": ["monsterId", "localeId"],
      "properties": {
        "monsterId": {
          "type": "string",
          "format": "guid"
        },
        "monster": {
          "$ref": "#/definitions/Monster"
        },
        "localeId": {
          "type": "string",
          "format": "guid"
        },
        "locale": {
          "$ref": "#/definitions/Locale"
        }
      }
    },
    "MonsterBuilding": {
      "type": "object",
      "required": ["monsterId", "buildingId"],
      "properties": {
        "monsterId": {
          "type": "string",
          "format": "guid"
        },
        "monster": {
          "$ref": "#/definitions/Monster"
        },
        "buildingId": {
          "type": "string",
          "format": "guid"
        },
        "building": {
          "$ref": "#/definitions/Building"
        }
      }
    },
...
    "BuildingMap": {
      "type": "object",
      "required": ["buildingId", "mapId"],
      "properties": {
        "buildingId": {
          "type": "string",
          "format": "guid"
        },
        "building": {
          "$ref": "#/definitions/Building"
        },
        "mapId": {
          "type": "string",
          "format": "guid"
        },
        "map": {
          "$ref": "#/definitions/Map"
        },
        "coords": {
          "type": "array",
          "items": {}
        }
      }
    },
    "JsonPatchDocumentOfBuilding": {
      "type": "object",
      "properties": {
        "operations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OperationOfBuilding"
          }
        },
        "contractResolver": {
          "$ref": "#/definitions/IContractResolver"
        }
      }
    },
    "OperationOfBuilding": {
      "allOf": [
        {
          "$ref": "#/definitions/Operation"
        },
        {
          "type": "object"
        }
      ]
    },
    "Operation": {
      "allOf": [
        {
          "$ref": "#/definitions/OperationBase"
        },
        {
          "type": "object",
          "properties": {
            "value": {}
          }
        }
      ]
    },
...
    "IContractResolver": {
      "type": "object",
      "x-abstract": true
    },
...
    "JsonPatchDocumentOfMonster": {
      "type": "object",
      "properties": {
        "operations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OperationOfMonster"
          }
        },
        "contractResolver": {
          "$ref": "#/definitions/IContractResolver"
        }
      }
    },
    "OperationOfMonster": {
      "allOf": [
        {
          "$ref": "#/definitions/Operation"
        },
        {
          "type": "object"
        }
      ]
    },
    "JsonPatchDocumentOfNpc": {
      "type": "object",
      "properties": {
        "operations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OperationOfNpc"
          }
        },
        "contractResolver": {
          "$ref": "#/definitions/IContractResolver"
        }
      }
    },
    "OperationOfNpc": {
      "allOf": [
        {
          "$ref": "#/definitions/Operation"
        },
        {
          "type": "object"
        }
      ]
    },
...
  }
}